refactor: 포인트객체 내부 타입 int -> Long 으로 변경

+++ src/main/java/cart/dao/PointDao.java
@@ -21,7 +21,7 @@ public class PointDao {
         this.jdbcTemplate = jdbcTemplate;
     }
 
-    public void update(final Long memberId, final int point) {
+    public void update(final Long memberId, final Long point) {
         String sql = "UPDATE point SET point = ? WHERE member_id = ?";
 
         final int affectedRow = jdbcTemplate.update(sql, point, memberId);
